requirements.txt
# requirements.txt
airtable-python-wrapper>=1.0.0
cryptography>=42.0.5
requests>=2.31.0
python-dotenv>=1.0.0